organized files, deleted old comments and tests that were not needed, refactored UVSim class into its own file and refactored test_sample into its own folder
