--- badtest1
@@ -1,17 +1,0 @@
-+1010
-+20+07
-
-20+09
--1110
-+1o0O
-3100+
--+1100
------
-+9999
--4300
--0000
-+4300
--0000
-+0000
-+0000
--99999
